Registration button and infos added
