+++ public/stylesheets/style-simplyk.css
@@ -51,7 +51,7 @@ body{
 	overflow: auto;
 }
 
-.conn-input{
+.tp-15{
 	padding-top: 15px;
 	padding-bottom: 15px;
 }
